Platform change for STM32U5 only: add I2C and SPI support.
This commit adds the I2C and SPI support that was missing in the previous commit.

Note that ST have acknowledged the problem with the ST-provided CMSIS layer for ThreadX not freeing memory when a dynamic task is terminated: they are looking into a fix, no timescale yet though; this is being progressed in https://github.com/STMicroelectronics/STM32CubeU5/issues/48.